+++ twake/frontend/src/app/services/channels/channels.js
@@ -2,7 +2,7 @@ import Observable from 'app/services/Depreciated/observable.js';
 import Workspaces from 'services/workspaces/workspaces.js';
 import DepreciatedCollections from 'app/services/Depreciated/Collections/Collections.js';
 import Collections from 'app/services/CollectionsReact/Collections';
-import LocalStorage from 'services/localStorage.js';
+import LocalStorage from 'app/services/LocalStorage';
 import WindowService from 'services/utils/window.js';
 import MenusManager from 'app/components/Menus/MenusManager.js';
 import Globals from 'services/Globals.js';
