+++ twake/frontend/src/app/services/login/login.js
@@ -16,6 +16,7 @@ import RouterServices from '../RouterService';
 import JWTStorage from 'services/JWTStorage';
 import AccessRightsService from 'services/AccessRightsService';
 import Environment from 'environment/environment';
+import LocalStorage from 'services/LocalStorage';
 
 class Login extends Observable {
   // Promise resolved when user is defined
@@ -305,7 +306,7 @@ class Login extends Observable {
 
   clearLogin() {
     this.currentUserId = null;
-    Globals.localStorageClear();
+    LocalStorage.clear();
     Collections.clear();
     JWTStorage.clear();
   }
